fix bug

+++ MIDP/loaders/msd_loader.py
@@ -85,7 +85,7 @@ class MSDLoader:
             data_idx + '.nii.gz'
         )).affine
         nib.save(
-            nib.Nifti1Image(prediction, affine=affine),
+            nib.Nifti1Image(prediction.astype(np.uint8), affine=affine),
             os.path.join(output_dir, data_idx + '.nii.gz')
         )
